backend/requirements.txt
# Core dependencies
flask
python-dotenv
openai
pydantic
echo "flask-cors"